Safari-611.1.10/LayoutTests/fast/canvas/canvas-strokeRect-expected.txt

Series of tests to ensure correct behaviour of canvas.strokeRect()

On success, you will see a series of "PASS" messages, followed by "TEST COMPLETE".


Test canvas.strokeRect() with solid green.
PASS imgdata[4] is 0
PASS imgdata[5] is 128
PASS imgdata[6] is 0
Test canvas.strokeRect() with a pattern.
PASS imgdata[4] is 0
PASS imgdata[5] is 128
PASS imgdata[6] is 0
Test canvas.strokeRect() with a gradient.
PASS imgdata[4] is 0
PASS imgdata[5] is 128
PASS imgdata[6] is 0
Test canvas.strokeRect() with height = width = 0 and lineWidth = 2.
PASS imgdata[0] is 0
PASS imgdata[1] is 0
PASS imgdata[2] is 0
Test canvas.strokeRect() with height = width = 0, lineWidth = 2, and shadow.
PASS imgdata[0] is 0
PASS imgdata[1] is 0
PASS imgdata[2] is 0
PASS successfullyParsed is true

TEST COMPLETE
